requirements.txt
python-dotenv==1.0.1
ics==0.7.2
opencc-python-reimplemented==0.1.7
openai>=1.47.0
watchdog==4.0.0
pytz==2024.2

